clean up App
