+++ src/Layout.css
@@ -1,0 +1,11 @@
+.header {
+  background-color: #222;
+  height: 50px;
+  padding: 20px;
+  color: white;
+  text-align: center;
+}
+
+.title {
+  font-size: 1.5em;
+}
